test(hash): move hash tests to integration target
